Set backup root directory as home of borgbackup user
This ensures internal cache and config data of borg is in the same place
as the explicit configuration. If the backup root is encrypted the data
in the home directory will be encrypted as well.

This also allows putting the SSH private key into the ~/.ssh directory
without the risk of storing it in a less sercurity place.