Set syncthing root directory as home of syncthing user
This ensures internal cache and config data of syncthing is in the same place
as the explicit configuration. If the root directory is encrypted the data
in the home directory will be encrypted as well.

Until now we created a syncthing symbolic link in the ~/.config directory
pointing to the .config folder in the syncthing root directory.